+++ android/app/.settings/org.eclipse.buildship.core.prefs
@@ -1,2 +1,2 @@
-#Wed Mar 14 16:45:24 JST 2018
+#Sat Mar 17 17:51:16 JST 2018
 connection.project.dir=..
