+++ src/Camera.js
@@ -3,41 +3,70 @@ import {
   Text,
   View,
   TouchableOpacity,
+  StyleSheet
 } from 'react-native';
 import { RNCamera } from 'react-native-camera';
-import styles from './styles/styles';
 
-const Camera = ({ navigation }) => {
-  const takePicture = async () => {
+class CameraScreen extends React.Component {
+  static navigationOptions = {
+    title: 'Take a picture of your product',
+  };
+  
+  takePicture = async () => {
     if (this.camera) {
       const options = { quality: 0.5, base64: true };
       const data = await this.camera.takePictureAsync(options);
-      navigation.navigate('Details', {
+      this.props.navigation.navigate('Detail', {
         picture: data.base64,
       });
     }
   };
 
-  return (
-    <View style={styles.container}>
-      <RNCamera
-        ref={(ref) => { this.camera = ref; }}
-        style={styles.preview}
-        type={RNCamera.Constants.Type.back}
-        flashMode={RNCamera.Constants.FlashMode.on}
-        permissionDialogTitle="Permission to use camera"
-        permissionDialogMessage="We need your permission to use your camera phone"
-      />
-      <View style={{ flex: 0, flexDirection: 'row', justifyContent: 'center' }}>
-        <TouchableOpacity
-          onPress={takePicture}
-          style={styles.capture}
-        >
-          <Text style={{ fontSize: 14 }}> SNAP </Text>
-        </TouchableOpacity>
+  render() {
+    const { navigate } = this.props.navigation;
+    return (
+      <View style={styles.container}>
+        <RNCamera
+          ref={(ref) => { this.camera = ref; }}
+          style={styles.preview}
+          type={RNCamera.Constants.Type.back}
+          flashMode={RNCamera.Constants.FlashMode.on}
+          permissionDialogTitle="Permission to use camera"
+          permissionDialogMessage="We need your permission to use your camera phone"
+        />
+        <View style={{ flex: 0, flexDirection: 'row', justifyContent: 'center' }}>
+          <TouchableOpacity
+            onPress={this.takePicture}
+            style={styles.capture}
+          >
+            <Text style={{ fontSize: 14 }}> SNAP </Text>
+          </TouchableOpacity>
+        </View>
       </View>
-    </View>
-  );
+    );  
+  }
 };
 
-export default Camera;
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    flexDirection: 'column',
+    backgroundColor: 'black',
+  },
+  preview: {
+    flex: 1,
+    justifyContent: 'flex-end',
+    alignItems: 'center',
+  },
+  capture: {
+    flex: 0,
+    backgroundColor: '#fff',
+    borderRadius: 5,
+    padding: 15,
+    paddingHorizontal: 20,
+    alignSelf: 'center',
+    margin: 20,
+  },
+});
+
+export default CameraScreen;
